Cclib

Documentation for Cclib.

Cclib.bridge

Cclib.cclib

Cclib.io

Cclib.method

Cclib.parser

Cclib.progress

Cclib.test

Cclib.bridge — Constant

bridge

Facilities for moving parsed data to other cheminformatic libraries.

Cclib.cclib — Constant

cclib

Variable from pyimport("cclib"). Numerous methods are available through Python's dot-call syntax.

Cclib.io — Constant

io

Available writers for standard chemical representations.

Cclib.method — Constant

method

Analyses and calculations based on data parsed by cclib.

Cclib.parser — Constant

parser

Parsers for all supported programs.

Cclib.progress — Constant

progress

Ways of iIndicating progress.

Cclib.test — Constant

test

Available writers for standard chemical representations.